requirements.txt
Django==5.1.2
dj-database-url==2.2.0
gunicorn==23.0.0
Pillow==10.4.0
qrcode==7.4.2
reportlab==4.2.2
whitenoise==6.7.0
psycopg[binary]>=3.2,<3.3
pywebview==4.4
